Add files via upload
Please note what the previous grade was and the reason for it. I added an input that lets you choose the amount of cities, its not because the code is not good... this is because api at the time of testing was taking HOURS to process many cities. Heres updated repo with 500+ cities after letting it run for hours.